move create user handler to login module
