+++ nodeAppPostPg/index.ts
@@ -35,7 +35,7 @@ import { RouteNamesSelfServiceInfoExists } from "./RouteNames";
 import { RouteNamesCancel, RouteNamesUnmatched,RouteNamesDetails  } from "./RouteNames";
 import { logging }          from "./logging";
 
-import { hashPassword, verifyUniqueUser, verifyCredentials } from './login';
+import { verifyUniqueUser, verifyCredentials, createUser } from './login';
 import { createToken } from './token';
 
 let dbQueriesPosts = new DbQueriesPosts();
@@ -287,48 +287,7 @@ server.route({
     pre: [
       {method: verifyUniqueUser}
     ],
-    handler: (req, res) => {
-
-      const errorMessage = "failed to add user";
-
-      // const payload = JSON.parse( req.payload.info);
-      const payload = req.query;
-    
-      let user = payload || {}
-    
-      const password = user.password;
-
-      hashPassword(password, async (err, hash) => {
-        if (err) {
-          console.log("bad info");
-
-          return res(Boom.badRequest(errorMessage));
-        }
-
-        user.isAdmin = user.isAdmin && user.isAdmin === "true" ? true : false;
-
-        // store info, and the hash rather than the pwd
-        user.password = hash;
-
-        console.log("user:", user);            
-
-        const uuid = await routeFns.addUserInternal(req, res, user);
-
-        if (!uuid) {
-          return res(Boom.badRequest(errorMessage));
-        }
-      
-        const token = createToken(user);
-
-        console.log("token:", token);
-
-        return res({ id_token: token}).code(201);
-      });
-    }
-    // ,
-    // validate: {
-    //   payload: createUserSchema
-    // }
+    handler: createUser
   }
 });
